+++ src/IO/MappingStream.cpp
@@ -391,7 +391,7 @@ namespace IO {
 
     void MappingStream::flush() {
         if (canWrite()) {
-            for (uint32_t i = 0; i < _views.count(); i++) {
+            for (size_t i = 0; i < _views.count(); i++) {
                 View *view = _views[i];
                 view->flush();
             }
