+++ src/communication/BaseCommService.cpp
@@ -215,7 +215,7 @@ namespace Communication {
         _instructionPool = nullptr;
 
         _loopSendersMutex.lock();
-        for (uint32_t i = 0; i < _loopSenders.count(); i++) {
+        for (size_t i = 0; i < _loopSenders.count(); i++) {
             BaseLoopSender *sender = _loopSenders[i];
             sender->stop();
         }
@@ -223,7 +223,7 @@ namespace Communication {
         _loopSendersMutex.unlock();
 
         _packetSendersMutex.lock();
-        for (uint32_t i = 0; i < _packetSenders.count(); i++) {
+        for (size_t i = 0; i < _packetSenders.count(); i++) {
             BasePacketSender *sender = _packetSenders[i];
             sender->stop();
         }
