+++ src/data/Convert.cpp
@@ -164,12 +164,12 @@ namespace Data {
         off_t offset = 0;
         int range = 0;
         char prevCh = '\0';
-        for (uint32_t i = 0; i < str.length(); i++) {
+        for (size_t i = 0; i < str.length(); i++) {
             char ch = str[i];
             if (ch == splitSymbol && prevCh != escape) {
                 if (range == 0) {
                     texts.add(str.substr(offset, i - offset));
-                    offset = i + 1;     // skip ;
+                    offset = (off_t) i + 1;     // skip ;
                 }
             } else if (ch == startRange && prevCh != escape) {
                 range++;
@@ -191,7 +191,7 @@ namespace Data {
         StringArray texts;
         Convert::splitItems(str, texts, splitSymbol, escape, startRange, endRange);
         if (texts.count() > 0) {
-            for (uint32_t i = 0; i < texts.count(); i++) {
+            for (size_t i = 0; i < texts.count(); i++) {
                 StringArray values;
                 Convert::splitItems(texts[i], values, ':');
                 if (values.count() == 2) {
