+++ src/driver/DriverDescription.cpp
@@ -1,14 +1,7 @@
 #include "exception/Exception.h"
-#include "data/DateTime.h"
-#include "diag/Trace.h"
-#include "diag/Trace.h"
-#include "thread/Thread.h"
-#include "thread/Locker.h"
-#include "thread/ThreadPool.h"
 #include "driver/devices/DeviceDescription.h"
 #include "driver/channels/ChannelDescription.h"
 #include "driver/DriverManager.h"
-#include "driver/devices/Sampler.h"
 
 using namespace Data;
 
@@ -33,7 +26,7 @@ namespace Drivers
     }
     const DeviceDescription* DriverDescription::getDevice(const String& name) const
     {
-        for (uint32_t i=0; i<_devices->count(); i++)
+        for (size_t i=0; i<_devices->count(); i++)
         {
             const DeviceDescription* dd = _devices->at(i);
             if(dd->name() == name)
